requirements.txt
pyqt5
opencv-python
pillow
pytesseract
reportlab
numpy
